cpp:S2275 — format specifier/argument type mismatch
The factory-reset Serial.printf uses `%lu` (unsigned long) but passes
Constants::FACTORY_RESET_HOLD_MS which is uint32_t. Wrap the argument in
static_cast<unsigned long> to match the format specifier.

On ESP32 (XTensa LX6, ILP32) int and long are both 32-bit, so the cast is
value-preserving; behaviour identical. Unchanged firmware binary size
(938617 bytes).

+++ firmware/can-bridge/can-bridge.ino
@@ -1276,7 +1276,7 @@ static bool checkFactoryReset() {
     // Check if button is pressed (low = pressed on GPIO0 with pullup)
     if (digitalRead(Constants::FACTORY_RESET_PIN) == LOW) {
         Serial.printf("%sFactory reset: GPIO0 held at boot, waiting %lums to confirm...%s\r\n",
-                        YELLOW, Constants::FACTORY_RESET_HOLD_MS, NC);
+                        YELLOW, static_cast<unsigned long>(Constants::FACTORY_RESET_HOLD_MS), NC);
 
         // Wait to see if button continues to be held
         uint32_t heldMs = 0;
